+++ core/controller/cluster/controller_basic.go
@@ -661,7 +661,7 @@ func (c *controller) GetClusterByName(ctx context.Context,
 
 // DeleteCluster TODO(gjq): failed to delete cluster, give user a alert.
 // TODO(gjq): add a deleting tag for cluster
-func (c *controller) DeleteCluster(ctx context.Context, clusterID uint) (err error) {
+func (c *controller) DeleteCluster(ctx context.Context, clusterID uint, hard bool) (err error) {
 	const op = "cluster controller: delete cluster"
 	defer wlog.Start(ctx, op).StopPrint()
 
@@ -734,9 +734,26 @@ func (c *controller) DeleteCluster(ctx context.Context, clusterID uint) (err err
 			log.Errorf(newctx, "failed to delete harbor repository: %v, err: %v", cluster.Name, err)
 		}
 
-		// 3. delete cluster in git repo
-		if err = c.clusterGitRepo.DeleteCluster(newctx, application.Name, cluster.Name, cluster.ID); err != nil {
-			log.Errorf(newctx, "failed to delete cluster: %v in git repo, err: %v", cluster.Name, err)
+		if hard {
+			if err = c.clusterGitRepo.HardDeleteCluster(newctx, application.Name, cluster.Name, cluster.ID); err != nil {
+				log.Errorf(newctx, "failed to delete cluster: %v in git repo, err: %v", cluster.Name, err)
+			}
+
+			if err := c.pipelinerunMgr.DeleteByClusterID(ctx, clusterID); err != nil {
+				log.Errorf(newctx, "failed to delete pipelineruns of cluster: %v, err: %v", cluster.Name, err)
+			}
+
+			if err := c.pipelineMgr.DeleteByClusterName(ctx, cluster.Name); err != nil {
+				log.Errorf(newctx, "failed to delete pipelineruns of cluster: %v, err: %v", cluster.Name, err)
+			}
+
+			if err := c.tagMgr.UpsertByResourceTypeID(ctx, common.ResourceCluster, clusterID, nil); err != nil {
+				log.Errorf(newctx, "failed to delete tags of cluster: %v, err: %v", cluster.Name, err)
+			}
+		} else {
+			if err = c.clusterGitRepo.DeleteCluster(newctx, application.Name, cluster.Name, cluster.ID); err != nil {
+				log.Errorf(newctx, "failed to delete cluster: %v in git repo, err: %v", cluster.Name, err)
+			}
 		}
 
 		// 4. delete cluster in db
